v2.0.0-major: Added textured model support
ModelTest executable can be ran to get a preview of a helmet model.
Supported texture types:
- Diffuse
- Normal
- Specular
- Metallic (Not present in demo)
- Roughness (Not present in demo)
- Ambient Occlusion
- Emissive

My immediate next plan at the moment is to figure out a robust and user-friendly mechanism for handling combined textures such as roughness-metallic textures which are very common. My current thoughts are on whether this should be the responsibility of NekiVK, though since the model loading process is pretty much entirely abstracted away from the user, they should be offered some mechanism to at least get the channels that specified data presides in.