v3.1.1-patch: Diagnosed swapchain destruction segfault issue
The issue was a result of glfwTerminate() - which destroys the VkSurfaceKHR - being run before the swapchain was destroyed. In the case of ModelTest.cpp::main(), this was solved by surrounding the ModelTest object in braces to control its lifetime. This also means that it becomes the burden of the user to correctly set up an appropriate-to-their-project solution to this.